+++ CHANGELOG.md
@@ -10,6 +10,16 @@
 ### Security
 
 
+## [2.2.0]
+### Added
+- add `ixf_last_import` to InternetExchangeBase (peeringdb/peeringdb#683)
+- add `ixf_net_count` to InternetExchangeBase (peeringdb/peeringdb#683)
+- add `ixf_ixp_member_list_url` to IXLanBase (peeringdb/peeringdb#249)
+- add `ixf_ixp_member_list_url_visible` to IXLanBase (peeringdb/peeringdb#249)
+### Fixed
+- fix issue in setup.py for test requirements
+
+
 ## [2.1.1]
 ### Fixed
 - remove verbose_name migrations (#45)